backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.28
pydantic>=2.10
pydantic-settings>=2.7
python-dotenv>=1.1
google-genai>=1.5
